rspy::enumerate をファイル分割
